# src/main/java/school/hei/federationagricole/entity/account/CashAccount.java
package school.hei.federationagricole.entity.account;

import lombok.Data;
import lombok.EqualsAndHashCode;
import lombok.NoArgsConstructor;
import lombok.experimental.SuperBuilder;

@Data
@SuperBuilder
@NoArgsConstructor
@EqualsAndHashCode(callSuper = true)
public class CashAccount extends FinancialAccount {
}